+++ efx/src/render.rs
@@ -1,10 +1,4 @@
-use crate::tags::button::render_button;
-use crate::tags::column::render_column_stmt;
-use crate::tags::hyperlink::render_hyperlink_stmt;
-use crate::tags::label::render_label_stmt;
-use crate::tags::row::render_row_stmt;
-use crate::tags::separator::render_separator_stmt;
-use crate::tags::text_field::render_text_field_stmt;
+use crate::tags::*;
 use efx_core::{Element, Node};
 use quote::{ToTokens, quote};
 
@@ -19,7 +13,7 @@ pub(crate) fn render_nodes_as_stmts<UI: ToTokens>(
     out
 }
 
-fn render_node_stmt<UI: ToTokens>(ui: &UI, node: &Node) -> proc_macro2::TokenStream {
+pub(crate) fn render_node_stmt<UI: ToTokens>(ui: &UI, node: &Node) -> proc_macro2::TokenStream {
     use efx_core::Node::*;
     match node {
         Text(t) => {
@@ -58,6 +52,7 @@ fn render_element_stmt<UI: ToTokens>(ui: &UI, el: &Element) -> proc_macro2::Toke
             quote! { #ts; }
         }
         "TextField" => render_text_field_stmt(ui, el),
+        "CentralPanel" => render_central_panel_stmt(ui, el),
         other => {
             let msg = format!("efx: unknown tag <{}>", other);
             quote! { compile_error!(#msg); }