+++ efx/Changelog.md
@@ -5,7 +5,7 @@
 - Label, Button, Row, Column, Separator attributes
 - Updated documentation (attributes, quickstarts for eframe/bevy/winit+wgpu)
 - New Tags: Heading, Hyperlink, TextField, Image, Grid
-- Added Panel Tags: Window, CentralPanel, SidePanel
+- Added Panel Tags: Window, CentralPanel, SidePanel, ScrollArea
 - Added examples
 - Added tests